Rename _foundation to _util
Let's drop this nasty '_foundation' legacy, which is only
source of bad memories
